requirements.txt
pytest~=8.4.0
selenium~=4.33.0
requests~=2.32.4
Faker~=37.4.0
allure-python-commons~=2.14.3
idna~=3.10